+++ web/source/osk/banner.ts
@@ -87,6 +87,24 @@ namespace com.keyman.osk {
     public getDiv(): HTMLElement {
       return this.div;
     }
+
+    /**
+     * Function     activate
+     * Scope        Public
+     * Description  Adds any relevant event listeners needed by this banner type.
+     */
+    public activate() {
+      // Default implementation - no listeners.
+    }
+
+    /**
+     * Function     activate
+     * Scope        Public
+     * Description  Removes any relevant event listeners previously added by this banner.
+     */
+    public deactivate() {
+      // Default implementation - no listeners.
+    }
   }
 
   /**
@@ -119,12 +137,14 @@ namespace com.keyman.osk {
         super(0);
       }
 
+      console.log("Loading img with src '" + imagePath + "'");
       this.img = document.createElement('img');
       this.img.setAttribute('src', imagePath);
       let ds = this.img.style;
       ds.width = '100%';
       ds.height = '100%';
       this.getDiv().appendChild(this.img);
+      console.log("Image loaded.");
     }
 
     /**
@@ -145,7 +165,7 @@ namespace com.keyman.osk {
     private display: HTMLSpanElement;
     private fontFamily?: string;
 
-    private suggestion: Suggestion;
+    private _suggestion: Suggestion;
 
     private index: number;
 
@@ -195,6 +215,10 @@ namespace com.keyman.osk {
       this.div['suggestion'] = this;
     }
 
+    get suggestion(): Suggestion {
+      return this._suggestion;
+    }
+
     /**
      * Function update
      * @param {string}     id           Element ID for the suggestion span
@@ -202,7 +226,7 @@ namespace com.keyman.osk {
      * Description  Update the ID and text of the BannerSuggestionSpec
      */
     public update(suggestion: Suggestion) {
-      this.suggestion = suggestion;
+      this._suggestion = suggestion;
       this.updateText();
     }
 
@@ -217,18 +241,18 @@ namespace com.keyman.osk {
      * @param target (Optional) The OutputTarget to which the `Suggestion` ought be applied.
      * Description  Applies the predictive `Suggestion` represented by this `BannerSuggestion`.
      */
-    public apply(target?: text.OutputTarget) {
+    public apply(target?: text.OutputTarget): text.Transcription {
       let keyman = com.keyman.singleton;
 
       if(this.isEmpty()) {
-        return;
+        return null;
       }
       
       // Find the state of the context at the time the prediction-triggering keystroke was applied.
-      let original = keyman.modelManager.getPredictionState(this.suggestion.transformId);
+      let original = keyman.modelManager.getPredictionState(this._suggestion.transformId);
       if(!original) {
         console.warn("Could not apply the Suggestion!");
-        return;
+        return null;
       } else {
         if(!target) {
           /* Assume it's the currently-active `OutputTarget`.  We should probably invalidate 
@@ -242,7 +266,7 @@ namespace com.keyman.osk {
 
         // Step 1:  determine the final output text
         let final = text.Mock.from(original.preInput);
-        final.apply(this.suggestion.transform);
+        final.apply(this._suggestion.transform);
 
         // Step 2:  build a final, master Transform that will produce the desired results from the CURRENT state.
         // In embedded mode, both Android and iOS are best served by calculating this transform and applying its
@@ -254,11 +278,16 @@ namespace com.keyman.osk {
         if(keyman['oninserttext'] && keyman.isEmbedded) {
           keyman['oninserttext'](transform.deleteLeft, transform.insert, transform.deleteRight);
         }
+
+        // Build a 'reversion' Transcription that can be used to undo this apply() if needed.
+        let preApply = text.Mock.from(original.preInput);
+        preApply.apply(original.transform);
+        return preApply.buildTranscriptionFrom(target, null);
       }
     }
 
     public isEmpty(): boolean {
-      return !this.suggestion;
+      return !this._suggestion;
     }
 
     /**
@@ -271,7 +300,7 @@ namespace com.keyman.osk {
       let keyman = com.keyman.singleton;
       let util = keyman.util;
 
-      let suggestion = this.suggestion;
+      let suggestion = this._suggestion;
       var suggestionText: string;
 
       var s=util._CreateElement('span');
@@ -301,7 +330,87 @@ namespace com.keyman.osk {
     }
   }
 
-  class SuggestionTouchManager extends dom.UITouchHandlerBase<HTMLDivElement> {
+  /**
+   * Function     SuggestionBanner
+   * Scope        Public
+   * @param {number} height - If provided, the height of the banner in pixels
+   * Description  Display lexical model suggestions in the banner
+   */
+  export class SuggestionBanner extends Banner {
+    public static readonly SUGGESTION_LIMIT: number = 3;
+    public static readonly MARGIN = 1;
+
+    private options : BannerSuggestion[];
+
+    private manager: SuggestionManager;
+
+    static readonly TOUCHED_CLASS: string = 'kmw-suggest-touched';
+
+    constructor(height?: number) {
+      super(height || Banner.DEFAULT_HEIGHT);
+
+      this.options = new Array();
+      for (var i=0; i<SuggestionBanner.SUGGESTION_LIMIT; i++) {
+        let d = new BannerSuggestion(i);
+        this.options[i] = d;
+        this.getDiv().appendChild(d.div);
+      }
+
+      this.manager = new SuggestionManager(this.getDiv(), this.options);
+
+      this.setupTouchHandling();
+    }
+
+    private setupTouchHandling() {
+      let keyman = com.keyman.singleton;
+      let div = this.getDiv();
+
+      let th = this.manager;
+
+      if(keyman.util.device.touchable) { //  /*&& ('ontouchstart' in window)*/ // Except Chrome emulation doesn't set this.
+        // Not to mention, it's rather redundant.
+        div.addEventListener('touchstart', function(e: TouchEvent) {
+          th.touchStart(e);
+        }, true);
+        // The listener below fails to capture when performing automated testing checks in Chrome emulation unless 'true'.
+        div.addEventListener('touchend', function(e: TouchEvent) {
+          th.touchEnd(e);
+        }, true); 
+        div.addEventListener('touchmove', function(e: TouchEvent) {
+          th.touchMove(e);
+        }, false);
+        //lDiv.addEventListener('touchcancel', osk.cancel,false); //event never generated by iOS
+      }
+    }
+
+    activate() {
+      let keyman = com.keyman.singleton;
+      let manager = this.manager;
+
+      keyman.modelManager['addEventListener']('invalidatesuggestions', manager.invalidateSuggestions);
+      keyman.modelManager['addEventListener']('suggestionsready', manager.updateSuggestions);
+      keyman.modelManager['addEventListener']('tryaccept', manager.tryAccept);
+      keyman.modelManager['addEventListener']('tryrevert', manager.tryRevert);
+
+      // Trigger a null-based initial prediction to kick things off.
+      keyman.modelManager.predict();
+    }
+
+    deactivate() {
+      let keyman = com.keyman.singleton;
+      let manager = this.manager;
+
+      keyman.modelManager['removeEventListener']('invalidatesuggestions', manager.invalidateSuggestions);
+      keyman.modelManager['removeEventListener']('suggestionsready', manager.updateSuggestions);
+      keyman.modelManager['removeEventListener']('tryaccept', manager.tryAccept);
+      keyman.modelManager['removeEventListener']('tryrevert', manager.tryRevert);
+    }
+  }
+
+  class SuggestionManager extends dom.UITouchHandlerBase<HTMLDivElement> {
+    private selected: BannerSuggestion;
+
+    //#region Touch handling implementation
     findTargetFrom(e: HTMLElement): HTMLDivElement {
       let keyman = com.keyman.singleton;
       let util = keyman.util;
@@ -330,7 +439,7 @@ namespace com.keyman.osk {
         console.warn("Cannot find BannerSuggestion object for element to highlight!");
       } else {
         // Never highlight an empty suggestion button.
-        let suggestion = t['suggestion'] as BannerSuggestion;
+        let suggestion = this.selected = t['suggestion'] as BannerSuggestion;
         if(suggestion.isEmpty()) {
           on = false;
         }
@@ -344,8 +453,7 @@ namespace com.keyman.osk {
     }
 
     protected select(t: HTMLDivElement): void {
-      let suggestion = t['suggestion'] as BannerSuggestion;
-      suggestion.apply();
+      this.doAccept(t['suggestion'] as BannerSuggestion);
     }
 
     //#region Long-press support
@@ -376,76 +484,165 @@ namespace com.keyman.osk {
       throw new Error("Method not implemented.");
     }
     //#endregion
+    //#endregion
 
-    constructor(div: HTMLElement) {
+    private options: BannerSuggestion[];
+
+    private initNewContext: boolean = true;
+
+    private currentSuggestions: Suggestion[] = [];
+    private currentTranscriptionID: number;
+
+    private recentAccept: boolean = false;
+    private recentAccepted: Suggestion;
+    private preAccept: text.Transcription = null;
+    private swallowPrediction: boolean = false;
+
+    private previousSuggestions: Suggestion[];
+    private previousTranscriptionID: number;
+
+    private recentRevert: boolean = false;
+    private rejectedSuggestions: Suggestion[] = [];
+
+    constructor(div: HTMLElement, options: BannerSuggestion[]) {
       // TODO:  Determine appropriate CSS styling names, etc.
       super(div, Banner.BANNER_CLASS, SuggestionBanner.TOUCHED_CLASS);
-    }
-  }
-
-  /**
-   * Function     SuggestionBanner
-   * Scope        Public
-   * @param {number} height - If provided, the height of the banner in pixels
-   * Description  Display lexical model suggestions in the banner
-   */
-  export class SuggestionBanner extends Banner {
-    public static readonly SUGGESTION_LIMIT: number = 3;
-    public static readonly MARGIN = 1;
-
-    private suggestionList : BannerSuggestion[];
-    private currentSuggestions: Suggestion[] = [];
-
-    private touchHandler: SuggestionTouchManager;
-
-    static readonly TOUCHED_CLASS: string = 'kmw-suggest-touched';
-
-    constructor(height?: number) {
-      super(height || SuggestionBanner.DEFAULT_HEIGHT);
-
-      this.suggestionList = new Array();
-      for (var i=0; i<SuggestionBanner.SUGGESTION_LIMIT; i++) {
-        let d = new BannerSuggestion(i);
-        this.suggestionList[i] = d;
-        this.getDiv().appendChild(d.div);
-      }
-
-      this.setupTouchHandling();
+      this.options = options;
     }
 
-    private setupTouchHandling() {
+    private doAccept(suggestion: BannerSuggestion) {
+      this.preAccept = suggestion.apply();
+      
+      this.selected = null;
+      this.recentAccept = true;
+      this.recentRevert = false;
+      this.recentAccepted = suggestion.suggestion;
+
+      this.previousSuggestions = this.currentSuggestions;
+      this.previousTranscriptionID = this.currentTranscriptionID;
+
+      // Request a 'new' prediction based on current context with a nil Transform.
       let keyman = com.keyman.singleton;
-      let div = this.getDiv();
-
-      let th = this.touchHandler = new SuggestionTouchManager(div);
-
-      if(keyman.util.device.touchable) { //  /*&& ('ontouchstart' in window)*/ // Except Chrome emulation doesn't set this.
-        // Not to mention, it's rather redundant.
-        div.addEventListener('touchstart', function(e: TouchEvent) {
-          th.touchStart(e);
-        }, true);
-        // The listener below fails to capture when performing automated testing checks in Chrome emulation unless 'true'.
-        div.addEventListener('touchend', function(e: TouchEvent) {
-          th.touchEnd(e);
-        }, true); 
-        div.addEventListener('touchmove', function(e: TouchEvent) {
-          th.touchMove(e);
-        }, false);
-        //lDiv.addEventListener('touchcancel', osk.cancel,false); //event never generated by iOS
-      }
+      this.swallowPrediction = true;
+      keyman.modelManager.predict();
     }
+
+    private doRevert() {
+      let keyman = com.keyman.singleton;
+      let current = text.Processor.getOutputTarget();
+      let priorState = this.preAccept;
+
+      // Step 1:  construct the reverted state.
+      let target = text.Mock.from(priorState.preInput);
+      target.apply(priorState.transform);
+
+      // Step 2:  build a final, master Transform that will produce the desired results from the CURRENT state.
+      // In embedded mode, both Android and iOS are best served by calculating this transform and applying its
+      // values as needed for use with their IME interfaces.
+      let transform = target.buildTransformFrom(current);
+      current.apply(transform);
+
+      // Signal the necessary text changes to the embedding app, if it exists.
+      if(keyman['oninserttext'] && keyman.isEmbedded) {
+        keyman['oninserttext'](transform.deleteLeft, transform.insert, transform.deleteRight);
+      }
+
+      // Denote the previous suggestion as rejected and update the 'valid' suggestion list accordingly.
+      this.rejectedSuggestions.push(this.recentAccepted);
+      this.currentSuggestions = this.previousSuggestions; // Restore to the previous state's Suggestion list.
+      this.currentTranscriptionID = this.previousTranscriptionID;
+
+      let rejectIndex = this.currentSuggestions.indexOf(this.recentAccepted);
+      if(rejectIndex != -1) {
+        this.currentSuggestions.splice(rejectIndex, 1);
+      }
+
+      // Other state maintenance
+      this.recentAccept = false;
+      this.recentRevert = true;
+      this.doUpdate();
+    }
+
+    /**
+     * Receives messages from the keyboard that the 'accept' keystroke has been entered.
+     * Should return 'false' if the current state allows accepting a suggestion and act accordingly.
+     * Otherwise, return true.
+     */
+    tryAccept: (source: string) => boolean = function(this: SuggestionManager, source: string): boolean {
+      if(!this.recentAccept && this.selected) {
+        this.doAccept(this.selected);
+        return false;
+      } else if(this.recentAccept && source == 'space') {
+        this.recentAccept = false;
+        return false; // Swallows a single space post-accept.
+      }
+      return true;  // Not yet implemented
+    }.bind(this);
+
+    /**
+     * Receives messages from the keyboard that the 'revert' keystroke has been entered.
+     * Should return 'false' if the current state allows reverting a recently-applied suggestion and act accordingly.
+     * Otherwise, return true.
+     */
+    tryRevert: () => boolean = function(this: SuggestionManager): boolean {
+      if(this.recentAccept) {
+        this.doRevert();
+        return false;
+      } else {
+        return true;
+      }
+    }.bind(this);
 
     /**
      * Function invalidateSuggestions
      * Scope        Public
      * Description  Clears the suggestions in the suggestion banner
      */
-    public invalidateSuggestions: (this: SuggestionBanner) => boolean = 
-        function(this: SuggestionBanner) {
-      this.suggestionList.forEach((option: BannerSuggestion) => {
+    public invalidateSuggestions: (this: SuggestionManager, source: text.prediction.InvalidateSourceEnum) => boolean = 
+        function(this: SuggestionManager, source: string) {
+
+      if(!this.swallowPrediction || source == 'context') {
+        this.recentAccept = false;
+        this.recentRevert = false;
+        this.rejectedSuggestions = [];
+
+        if(source == 'context') {
+          this.swallowPrediction = false;
+          this.initNewContext = true;
+        }
+      }
+
+      this.options.forEach((option: BannerSuggestion) => {
         option.update(null);
       });
     }.bind(this);
+
+    private doUpdate() {
+      let keyman = com.keyman.singleton;
+
+      let suggestions = [];
+      // TODO:  Insert 'current text' if/when valid as the leading option.
+      //        We need the LMLayer to tell us this somehow.
+      if(!this.recentAccept && !this.recentRevert && !this.initNewContext) {
+        // In the meantime, a placeholder:
+        // (generated from the 'true'/original transform)
+        let original = keyman.modelManager.getPredictionState(this.currentTranscriptionID);
+        let nil = new text.TextTransform(original.transform.insert + ' ', original.transform.deleteLeft, original.transform.deleteRight);
+
+        suggestions.push({transform: nil, displayAs: '&lt;keep&gt;', transformId: this.currentTranscriptionID});
+      }
+      suggestions = suggestions.concat(this.currentSuggestions);
+
+      this.options.forEach((option: BannerSuggestion, i: number) => {
+        if(i < suggestions.length) {
+          option.update(suggestions[i]);
+        } else {
+          option.update(null);
+        }
+      });
+
+      this.initNewContext = false;
+    }
 
     /**
      * Function updateSuggestions
@@ -453,17 +650,25 @@ namespace com.keyman.osk {
      * @param {Suggestion[]}  suggestions   Array of suggestions from the lexical model.
      * Description    Update the displayed suggestions in the SuggestionBanner
      */
-    public updateSuggestions: (this: SuggestionBanner, suggestions: Suggestion[]) => boolean =
-        function(this: SuggestionBanner, suggestions: Suggestion[]) {
-      this.currentSuggestions = suggestions;
+    public updateSuggestions: (this: SuggestionManager, prediction: text.prediction.ReadySuggestions) => boolean =
+        function(this: SuggestionManager, prediction: text.prediction.ReadySuggestions) {
       
-      this.suggestionList.forEach((option: BannerSuggestion, i: number) => {
-        if(i < suggestions.length) {
-          option.update(suggestions[i]);
-        } else {
-          option.update(null);
-        }
-      });
+      let suggestions = prediction.suggestions;
+
+      this.currentSuggestions = suggestions;
+      this.currentTranscriptionID = prediction.transcriptionID;
+
+      // If we've gotten an update request like this, it's almost always user-triggered and means the context has shifted.
+      if(!this.swallowPrediction) {
+        this.recentAccept = false;
+        this.recentRevert = false;
+        this.rejectedSuggestions = [];
+      } else { // This prediction was triggered by a recent 'accept.'  Now that it's fulfilled, we clear the flag.
+        this.swallowPrediction = false;
+      }
+
+      // The rest is the same, whether from input or from "self-updating" after a reversion to provide new suggestions.
+      this.doUpdate();
     }.bind(this);
   }
 }