+++ web/history.md
@@ -8,6 +8,7 @@
     * Centralizes code paths to improve parity between hardware and OSK-based keystrokes.
   * In both cases above, significant unit testing was facilitated and has been added as a result, further improving code maintainability into the future.
 * Began adding support for our common LMLayer interface for predictive modeling.
+* Fix for issue with Mandarin Chinese keyboard's 'picker' UI
 
 ## 2019-02-25 11.0.220 stable
 * 11.0 Stable release